requirements.txt
Pillow==10.3.0
opencv-python==4.10.0.84
numpy==1.26.4
python-doctr[torch]==0.8.1
torch==2.3.1
torchvision==0.18.1
transformers==4.45.1
python-magic==0.4.27
